requirements.txt
pytest
selenium
appium-python-client
